move existing error tests to a new file
